requirements.txt
python-telegram-bot==20.*  
python-dotenv  
sqlalchemy  
python-docx  
flask